MT#20543 Fix spelling unit test
Check manually for the Lancaster Consensus variables instead of using
the Test::DescribeMe module which is not packaged in Debian.

Conditionally use the Test::Spelling module instead of blindly assuming
it is there, as we only run the unit tests when doing author testing,
there is no point in always requiring the module.

Fix all_pod_files_spelling_ok() call, and remove done_testing which
should not be called in this case.

Add some new stop words.

Change-Id: I625cd104093f38cbd36a7439dbd8562d69638d55