MT#20543 Fix test suite
Pass a version option to strictures' import instead of a scalar. Enable
language features in the callers scope instead of in the Sipwise::Base
module scope. Protect the syntax that we are checking with eval, so that
when it is not available we can do proper test suite checks instead of
exiting with compilation errors. When testing unlink, use a filename
that is very unlikely to exist. Remove dh_auto_test override to enable
the test suite.

Change-Id: I458f680d4743ec9025ff68c44b867a7e6398fcd6