MT#20543 Fix test suite
Pass a version option to strictures' import instead of a scalar. Enable
language features in the callers scope instead of in the Sipwise::Base
module scope. Protect the syntax that we are checking with eval, so that
when it is not available we can do proper test suite checks instead of
exiting with compilation errors. Remove check for switch feature, which
is considered experimental with latest perl versions and emits warnings.
When testing unlink, use a filename that is very unlikely to exist.
Add libipc-system-simple-perl to Build-Depends required for
Fatalized/autodying system(), and libmodule-build-perl which will be
removed from perl's core. Remove dh_auto_test override to enable the
test suite.

Change-Id: I458f680d4743ec9025ff68c44b867a7e6398fcd6