Update patch set 1
Patch Set 1:

> actually this is not a fix, but quite a radical change.

Hmm, yeah I also pondered about this, and I guess it depends on what you consider the intended canonical behavior, the one assumed by current code users, or from the test suite. The change makes sense when trying to make the test suite work. The problem here indeed is that because the test suite has never worked it has not much standing as any kind of reference.

 >  Are you sure, all the code is ready for this?

I've rechecked what it enabled and I think 'say', 'state', 'switch' and 'array_base' are pretty safe, the former three because otherwise they cannot be used and would fail to compile, and the latter because it used to be the default, and it just stopped being provided with later versions. But 'unicode_strings' is not. So I'm tempted to either use :5.10 (and disable the unicode unit test), or consider this change out of scope and disable all the enabled features in the test suite, and propose it as part of another bug, alongside Rene's thread unit test for example, and other improvements.

Patch-set: 1